+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## 20.12 - 2020-12-
 
+### Breaking changes
+
+Options `--disable-pipes` and `--enable-pipes` (TeX engines) are deprecated and will be removed in a future release. If specified, they now don't have any effect.
+
 ### Fixed issues
 
 - [597](https://github.com/MiKTeX/miktex/issues/597): xetex rounds all image sizes to integer pt values
@@ -9,6 +13,7 @@
 - [659](https://github.com/MiKTeX/miktex/issues/659): Check for directory that executable doesn't have permissions for causes crash
 - [666](https://github.com/MiKTeX/miktex/issues/666): miktex (lualatex.exe) fails to run from flask subprocess
 - [668](https://github.com/MiKTeX/miktex/issues/668): expand tilde prefixes
+- [669](https://github.com/MiKTeX/miktex/issues/669): testing if pipes are enabled.
 
 ## 20.11 - 2020-11-02